requirements.txt
Django==1.10
djangorestframework==3.4.6
selenium==2.53.6
wheel==0.24.0
